AddTeamToRepoCommand with BaseCommand
