+++ src/Octoshift/Commands/ICommandArgs.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OctoshiftCLI.Commands
+{
+    public interface ICommandArgs
+    {
+    }
+}
